move test for TS DFC Catalogs in a separate modules, plus minor fixes
